Correcciones de CheckStyle

+++ src/main/java/ar/edu/utn/frba/dds/simeal/models/entities/colaboraciones/donarDinero/DonarDinero.java
@@ -2,11 +2,11 @@ package ar.edu.utn.frba.dds.simeal.models.entities.colaboraciones.donarDinero;
 
 import ar.edu.utn.frba.dds.simeal.models.entities.colaboraciones.Colaboracion;
 import ar.edu.utn.frba.dds.simeal.models.entities.personas.Colaborador;
+import java.time.LocalDate;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
-import java.time.LocalDate;
 
 @Getter
 @Builder
@@ -15,14 +15,14 @@ public class DonarDinero implements Colaboracion {
   private Colaborador colaborador;
   private Frecuencia frecuencia;
   private LocalDate fechaDeRealizacion;
-  private Integer cantidadDinero;
+  private double cantidadDinero;
   @Builder.Default
   private double factorDeReconocimiento = 0.5;
 
 
 
   @Override
-  public double calcularReconocimientoParcial(){
+  public double calcularReconocimientoParcial() {
     return cantidadDinero * this.factorDeReconocimiento;
   }
 }
